requirements.txt
fastapi==0.95.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==1.10.21
python-multipart==0.0.9
firebase-admin==6.4.0
openai==1.12.0
google-generativeai==0.8.4
httpx==0.27.0
tenacity==8.2.3
aiohttp==3.9.3
pytest==8.0.2
pytest-asyncio==0.23.5 